consolidate duplicate named files in util
